separate ZMUser's NSString extension function into a new file
